+++ app/src/main/kotlin/com/architects/pokearch/ui/features/details/extensions/getGradientImageAndColors.kt
@@ -1,0 +1,36 @@
+package com.architects.pokearch.ui.features.details.extensions
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import coil.compose.AsyncImagePainter
+import coil.compose.rememberAsyncImagePainter
+import com.architects.pokearch.domain.model.PokemonInfo
+import com.architects.pokearch.ui.components.extensions.GetColorsBackground
+import com.architects.pokearch.ui.components.extensions.buildImageRequest
+
+@Composable
+fun getGradientImageAndColors(
+    pokemonInfo: PokemonInfo,
+): Pair<AsyncImagePainter, List<Color>> {
+
+    val image = rememberAsyncImagePainter(
+        model = pokemonInfo.getOfficialArtworkImageUrl().buildImageRequest(LocalContext.current)
+    )
+
+    val colorDefault = MaterialTheme.colorScheme.surfaceVariant
+    val colorsDefault = listOf(colorDefault, colorDefault)
+
+    var colors by remember { mutableStateOf(colorsDefault) }
+
+    image.GetColorsBackground {
+        colors = it.ifEmpty { colorsDefault }
+    }
+
+    return (Pair(image, colors))
+}
